+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2021-02-09
+
+### Changed
+- [PG-1313](https://bighealth.atlassian.net/browse/PG-1313) Remove the online status ([#14](https://github.com/sleepio/community-app-web/pull/14))
+
 ## 2021-02-06
 
 ### Changed